Started implementing the CMakeLists file.
Currently very incomplete and still not working. I'm going to finish
writing this in the coming week, and try to get at least the VS MSVC
target to work as expected. We also need to maybe export or create a
FindPackage module for libSPEAR too. But that might be in the future.