Bundle SPIRV and SPIRV-Tools headers
Different SDKs bundle the spirv headers in different location, e.g.
LunarG SDK bundles them in the vulkan path, other do it in the spirv
path. Officially, Khronos puts them in the spirv path (see the repo).
The hack to fix this is to simply use the Khronos repo and submodule
this. This fixes all problems and also removed an implicit dependency
on having the spirv and spirv-tools modules somewhere in the system.
However, we still assume the user has the Vulkan headers somewhere, but
that's find, since CMake has built-in functionality for dealing with it.

TODO: make things actually build for MSVC and the VS target. After that,
make sure CMake exports things right. Finally, try to make it work for
the other targets, gcc and clang on Linux/BSD. Might need to fix some
things in the HLX submodule as we go along, as well as in the libSPEAR.